Fix reading parents of nested structs in arrays
0.9.0 introduced passing parents as objects in the second parameter to a struct reader,
but not everything was updated to match that. In particular, offsets would not be tracked
correctly in nested structs in StructTypes that don't change the parent object (array, if, …).
This would cause reading to continue at the start of the nested struct, rather than after the
nested struct, when returning from one.