Add object path to error message if reading fails (#33)
Currently, it can be really hard to figure out where reading failed—you get an error from deep inside `Struct.types.int32`, for example, but if there's hundreds of int32s in  your structure that doesn't really help.
This patch adds the path of the thing that could not be read to the error message. Performance seems to be barely affected.